+++ Runtime/Scripts/PubSub/Messenger.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using PJ.Native.Proto;
 using UnityEngine;
 
 namespace PJ.Native.PubSub
@@ -9,18 +10,18 @@ namespace PJ.Native.PubSub
     public sealed class Messenger : ReceivablePublisher
     {
         private Dictionary<string, Action<MessageHolder>> handlerMap;
-        private List<(Action<MessageHolder>, Predicate<MessageHolder>)> conditionHandlers;
+        private List<(Action<MessageHolder>, Predicate<Message>)> conditionHandlers;
 
         public Messenger(Tag tag) : base(tag)
         {
             MessageManager.Instance.Mediator.Register(this);
             handlerMap = new Dictionary<string, Action<MessageHolder>>();
-            conditionHandlers = new List<(Action<MessageHolder>, Predicate<MessageHolder>)>();
+            conditionHandlers = new List<(Action<MessageHolder>, Predicate<Message>)>();
         }
 
         public override bool HasKey(string key)
         {
-            return handlerMap.  ContainsKey(key);
+            return handlerMap.ContainsKey(key);
         }
 
         public override void OnReceive(MessageHolder messageHolder)
@@ -29,11 +30,11 @@ namespace PJ.Native.PubSub
             {
                 handler.Invoke(messageHolder);
             }
-            foreach(var conditionHandler in conditionHandlers)
+            foreach((Action<MessageHolder> callback, Predicate<Message> condition) in conditionHandlers)
             {
-                if(conditionHandler.Item2.Invoke(messageHolder))
+                if(condition.Invoke(messageHolder.Message))
                 {
-                    conditionHandler.Item1.Invoke(messageHolder);
+                    callback.Invoke(messageHolder);
                 }
             }
         }
@@ -44,17 +45,17 @@ namespace PJ.Native.PubSub
 
         }
 
-        public void UnSubscribe(string key)
+        public void Unsubscribe(string key)
         {
             handlerMap.Remove(key);
         }
 
-        public void Subscribe(Action<MessageHolder> handler, Predicate<MessageHolder> condition)
+        public void Subscribe(Action<MessageHolder> handler, Predicate<Message> condition)
         {
             conditionHandlers.Add((handler, condition));
         }
 
-        public void UnSubscribe(Action<MessageHolder> handler)
+        public void Unsubscribe(Action<MessageHolder> handler)
         {
             conditionHandlers.RemoveAll(conditionHandler => conditionHandler.Item1 == handler);
         }