+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased]
 ### Changed
+- Renmae unity package NativeBridge to NativePubSub
 - Rename namespace PJ.Native.Messenger to PJ.Native.PubSub
+- Rename Folder Runtime/Scripts/Messenger to Runtime/Scripts/PubSub
 - Rename MessageHandler's method
 - Rename MessageHandler to Messenger.
 ### Removed